LayoutTests/media/video-append-source.html

<video controls></video>
<script src=video-test.js></script>

<script>
    testExpected("video.currentSrc", "");
    var source = document.createElement("source");
    source.setAttribute("src", "content/test.mp4");
    video.appendChild(source);

    testExpected("video.currentSrc", "");

    waitForEvent("load", function () {
        testExpected("relativeURL(video.currentSrc) ", "content/test.mp4");
        endTest();
    });

</script>
