LayoutTests/media/video-append-source-expected.txt

EXPECTED (video.currentSrc == '') OK
EXPECTED (video.currentSrc == '') OK
EVENT(load)
EXPECTED (relativeURL(video.currentSrc) == 'content/test.mp4') OK
END OF TEST
